requirements.txt
matplotlib==3.2.2  
numpy==1.18.5  
scipy==1.4.1  
keras==2.4.3  
tensorflow==2.3.0  
torch==1.7.0  
Pillow==7.0.0  
scikit-image==0.16.2  
scikit-learn==0.22.2
